ajj ki dusriicommit

+++ controller/all.js
@@ -1,46 +1,75 @@
 const create = require("../models/create")
+const task = require("../models/task")
+const mytask =require("../models/mytask")
+
+// exports.PostHandler = async(req,res)=>{
+// try{
+// const{title,body,description}= req.body
+// console.log(req.body)
+// if(title =="" && body =="" && description =="" ){
+// return console.log("all fields must be present ")
+// }
+// const cratoo = create.create({
+//     title,
+//     body,
+//     description
+
+// })
+ 
+// if(cratoo){
+// res.json("Hurray Your Item is created")
+
+// }
+// else{
+//     res.json("something went wrong") 
+// }
+
+
+
+// }
+// catch(error){
+//     console.log("error",error)
+// }
+
+// }
 
 
 exports.PostHandler = async(req,res)=>{
-try{
-const{title,body,description}= req.body
-console.log(req.body)
-if(title =="" && body =="" && description =="" ){
-return console.log("all fields must be present ")
-}
-const cratoo = create.create({
-    title,
-    body,
-    description
-
-})
- 
-if(cratoo){
-res.json("Hurray Your Item is created")
-
-}
-else{
-    res.json("something went wrong") 
-}
-
-
-
-}
-catch(error){
-    console.log("error",error)
-}
-
-}
-
-
-
+    try{
+    const{title,starttime,endtime}= req.body
+    console.log(req.body)
+    
+    const cratoo = mytask.create({
+        title,
+        starttime,
+        endtime
+    
+    })
+     
+    if(cratoo){
+    res.json("Hurray Your task is added")
+    
+    }
+    else{
+        res.json("something went wrong") 
+    }
+    
+    
+    
+    }
+    catch(error){
+        console.log("error",error)
+    }
+    
+    }
+    
 
 
 
 exports.ReadHandler=(req,res)=>{
     try{
     const {postId} = req.params
-    const read = create.findOne(postId)
+    const read = task.findOne(postId)
     if(read){
     res.json("post ha chui yekya baya",read)
     }
@@ -63,15 +92,18 @@ exports.ReadHandler=(req,res)=>{
 
 
 
-   exports.GetallHandler = (req,res)=>{
+   exports.GetallHandler = async (req,res)=>{
 
 
-        const alls=create.find()
-         if(!alls){
-         res.json("atii na chui nea kehen baya")
         
-         }
-         res.json("yekya chui sorui khn bayaaa", alls)
+            const alls = await mytask.find().exec(); 
+            if (!alls || alls.length === 0) {
+                return res.status(404).json({ message: "No products found" });
+            }
+            
+            const results = alls.map(item => item.toObject());
+            res.status(200).json({ message: "Products fetched successfully", data: results });
+        
         
         
         
@@ -88,21 +120,24 @@ exports.ReadHandler=(req,res)=>{
 
 
 
-       exports.UpdateHandler = (req,res)=>{
+       exports.UpdateHandler =async (req,res)=>{
             try{
-                const {postId} = req.params
-            const {title,body,description}= req.body
+                
+                const{_id,title,starttime,endtime,completed}= req.body
+                console.log(req.body)
+               
             
-              if (!postId || !title || !body || !description) {
+              if (!_id||!title || !starttime || !endtime || !completed ){
                 return res.status(400).json({ message: "Required fields are missing" });
             }
-             const update= create.findByIdAndUpdate(postId,{
+             const update= await mytask.findByIdAndUpdate(_id,{
             title,
-            body,
-            description
+            starttime,
+            endtime,
+            completed
              })
              if(update){
-              res.json("update ha gov baya",update)
+              res.json("update ha gov baya")
             
              }
              else{
@@ -130,9 +165,14 @@ exports.ReadHandler=(req,res)=>{
 
 
            exports.DeleteHandler = async (req, res)=>{
-                const{postId} = req.params
+                const {postId} = req.query
+                console.log(req.query,"id")
+                if(!postId){
+                    return res.status(400).json({ message: "Required fields are missing" });
+                }
+                console.log(postId)
                 try{
-                const delid = create.findByIdAndDelete(postId)
+                const delid =await create.findByIdAndDelete(postId)
                 if(delid){
                     res.json("Post Deleted as per Your wish")
                 }
